Move classes that are still in use out of “legacy” package
