+++ src/main/java/uk/gov/justice/hmpps/casenotes/controllers/SystemGeneratedController.kt
@@ -21,12 +21,12 @@ import uk.gov.justice.hmpps.casenotes.config.AuthAwareAuthenticationToken
 import uk.gov.justice.hmpps.casenotes.config.CaseNoteRequestContext
 import uk.gov.justice.hmpps.casenotes.config.SYSTEM_GENERATED_OPERATIONS
 import uk.gov.justice.hmpps.casenotes.config.SecurityUserContext.Companion.ROLE_SYSTEM_GENERATED_RW
+import uk.gov.justice.hmpps.casenotes.events.pusher.CaseNoteEventPusher
 import uk.gov.justice.hmpps.casenotes.integrations.PrisonerSearchService
-import uk.gov.justice.hmpps.casenotes.legacy.dto.ErrorResponse
-import uk.gov.justice.hmpps.casenotes.legacy.service.CaseNoteEventPusher
 import uk.gov.justice.hmpps.casenotes.notes.CaseNote
 import uk.gov.justice.hmpps.casenotes.systemgenerated.CreateSysGenNote
 import uk.gov.justice.hmpps.casenotes.systemgenerated.SystemGeneratedRequest
+import uk.gov.justice.hmpps.casenotes.utils.ErrorResponse
 
 @Tag(name = SYSTEM_GENERATED_OPERATIONS)
 @RestController
