+++ src/main/java/uk/gov/justice/hmpps/casenotes/notes/WriteCaseNote.kt
@@ -19,7 +19,7 @@ import uk.gov.justice.hmpps.casenotes.events.PersonCaseNoteEvent.Companion.creat
 import uk.gov.justice.hmpps.casenotes.events.PersonCaseNoteEvent.Type.CREATED
 import uk.gov.justice.hmpps.casenotes.events.PersonCaseNoteEvent.Type.DELETED
 import uk.gov.justice.hmpps.casenotes.events.PersonCaseNoteEvent.Type.UPDATED
-import uk.gov.justice.hmpps.casenotes.legacy.service.EntityNotFoundException
+import uk.gov.justice.hmpps.casenotes.utils.EntityNotFoundException
 import java.util.UUID.fromString
 
 @Service
