+++ src/main/java/uk/gov/justice/hmpps/casenotes/utils/UserMdcFilter.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.hmpps.casenotes.legacy.utils;
+package uk.gov.justice.hmpps.casenotes.utils;
 
 import lombok.extern.slf4j.Slf4j;
 import org.slf4j.MDC;
